created /ping and /login api tests
